requirements.txt
Django==4.1.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-environ==0.9.0
requests==2.28.1
django-environ==0.9.0
